requirements.txt
streamlit==0.76.0
tensorflow==2.5.0
opencv-python==4.5.1.48
Pillow==8.0.1
numpy==1.19.5
